[Win32] Correct DPI context awareness value retrieval
The DPI_AWARENESS_CONTEXT values are currently defined as constants, but
they are actually defined as typedefed handles instead of compile time
constants in Windows' windef.h. So the constants defined in SWT may not
fit the actual values as used by the OS. This currently seems to be the
case for the System awareness value, which does not fit the OS value.

This change adapts the definition of the values to retrieve the actual
current values from the OS instead of hardcoding them in constants.

+++ bundles/org.eclipse.swt/Eclipse SWT/win32/org/eclipse/swt/widgets/Display.java
@@ -5408,7 +5408,7 @@ public boolean setRescalingAtRuntime(boolean activate) {
 }
 
 private boolean setMonitorSpecificScaling(boolean activate) {
-	int desiredApiAwareness = activate ? OS.DPI_AWARENESS_CONTEXT_PER_MONITOR_AWARE_V2 : OS.DPI_AWARENESS_CONTEXT_SYSTEM_AWARE;
+	long desiredApiAwareness = activate ? OS.DPI_AWARENESS_CONTEXT_PER_MONITOR_AWARE_V2 : OS.DPI_AWARENESS_CONTEXT_SYSTEM_AWARE;
 	if (Win32DPIUtils.setDPIAwareness(desiredApiAwareness)) {
 		rescalingAtRuntime = activate;
 		coordinateSystemMapper = activate ? new MultiZoomCoordinateSystemMapper(this, this::getMonitors) : new SingleZoomCoordinateSystemMapper(this);
